Update STOP REASON, fix update Back to LOG

+++ Andon.Biz.Logic/MES_WORK_ORDER.cs
@@ -23,6 +23,6 @@ namespace iAndon.Biz.Logic
         public System.DateTime DEADLINE { get; set; }
         public string DESCRIPTION { get; set; }
         public string PO_CODE { get; set; }
-        public int STATUS { get; set; }
+        public short STATUS { get; set; }
     }
 }
