+++ Andon.Biz.Logic/MES_WORK_PLAN.cs
@@ -18,7 +18,7 @@ namespace iAndon.Biz.Logic
         public string LINE_ID { get; set; }
         public decimal DAY { get; set; }
         public string SHIFT_ID { get; set; }
-        public int STATUS { get; set; }
+        public short STATUS { get; set; }
         public string WORKING_ID { get; set; }
     }
 }
