+++ Andon.Biz.Logic/MES_WORK_PLAN_DETAIL_HISTORY.cs
@@ -29,9 +29,9 @@ namespace iAndon.Biz.Logic
         public string PRODUCT_CODE { get; set; }
         public string CONFIG_ID { get; set; }
         public decimal TAKT_TIME { get; set; }
-        public int STATION_QUANTITY { get; set; }
-        public int BATCH { get; set; }
-        public int HEAD_COUNT { get; set; }
+        public short STATION_QUANTITY { get; set; }
+        public short BATCH { get; set; }
+        public short HEAD_COUNT { get; set; }
         public decimal PLAN_QUANTITY { get; set; }
         public string DESCRIPTION { get; set; }
         public string USER_ACTION { get; set; }
